\name{closeProject}
\alias{closeProject}
\title{Close Project}
\usage{
closeProject(simplace)
}
\arguments{
  \item{simplace}{handle to the SimplaceWrapper object
  returned by \code{\link{initSimplace}}}
}
\description{
Finalizes the project
}
\details{
Call to the finalize method of the simulation.
}
\seealso{
\code{\link{openProject}}
}

