\name{getSimulationIDs}
\alias{getSimulationIDs}
\title{List with the IDs of the performed simulations}
\usage{
  getSimulationIDs(simplace)
}
\arguments{
  \item{simplace}{handle to the SimplaceWrapper object}
}
\value{
  list with the IDs
}
\description{
  Returns a vector with the IDs of the simulations. IDs are
  required to get the output of the simulations.
}

