\name{runSimulations}
\alias{runSimulations}
\title{Run the simulations}
\usage{
  runSimulations(simplace, updateresources = FALSE,
    selectsimulation = FALSE)
}
\arguments{
  \item{simplace}{handle to the SimplaceWrapper object
  returned by \code{\link{initSimplace}}}

  \item{updateresources}{if true update ressources}

  \item{selectsimulation}{if true keeps a selected
  simulation}
}
\description{
  Run the created simulations.
}

