% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplaceWrapper.R
\name{setSimulationValues}
\alias{setSimulationValues}
\title{Changes values of the current simulation}
\usage{
setSimulationValues(simplace, parameterList = NULL)
}
\arguments{
\item{simplace}{handle to the SimplaceWrapper object returned by \code{\link{initSimplace}}}

\item{parameterList}{a list with the parameter name as key and parametervalue as value}
}
\description{
Sets values of arbitrary SimVariables in a simplace simulation.
Useful if you want to couple simplace with another simulation 
and interchange values daily.
}
\examples{
\dontrun{
for(i in 1:365)
{
  param <- list(iRain=0.0)
  setSimulationValue(simplace,param)
  step(simplace)
}
}

}
