% Generated by roxygen2 (4.0.0): do not edit by hand
\name{setSlotCount}
\alias{setSlotCount}
\title{Sets the number of processors that are used parallel}
\usage{
setSlotCount(count)
}
\arguments{
\item{count}{number of processors}
}
\description{
The function can be used only after \code{\link{initSimplace}} has been
called.
}

