% Generated by roxygen2 (4.0.0): do not edit by hand
\name{varmapToList}
\alias{varmapToList}
\title{Convert varmap to list}
\usage{
varmapToList(varmap, expand = TRUE)
}
\arguments{
\item{varmap}{the varMap returned by \code{\link{stepSimulation}}}

\item{expand}{if \code{TRUE} expand array objects to vector.}
}
\value{
list with parameter name as key and parameter value as value
}
\description{
Converts the varmap to a list
}
\details{
Converts the varMap to a list. All elements are converted to appropriate
R objects. Arrays are expanded to vectors by default.
}
\examples{
\dontrun{
simplace <- initSimplace(SimplaceInstallationDir,SimplaceWorkDir,SimplaceOutputDir)
openProject(simplace, Solution)
varmap <- stepSimulation(simplace,count=22)
closeProject(simplace)

varlist <- varmapToList(varmap)

varlist$startdate - 24*3600
varlist$LintulBiomass.sWSO}
}

