% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SimplaceWrapper.R
\name{initSimplace}
\alias{initSimplace}
\title{Initialisation of Framework}
\usage{
initSimplace(InstallationDir, WorkDir, OutputDir, additionalClasspaths = c(),
  javaparameters = getOption("java.parameters"), force.init = TRUE)
}
\arguments{
\item{InstallationDir}{directory where simplace, lap and simplacerun are located}

\item{WorkDir}{working directory where solutions, projects and data resides (_WORKDIR_)}

\item{OutputDir}{directory for output (_OUTPUTDIR_)}

\item{additionalClasspaths}{vector with class paths relative to InstallationDir that are to be added}

\item{javaparameters}{parameters that are passed to the java virtual machine}

\item{force.init}{(re)initialize a running JVM, see \code{\link{.jinit}}}
}
\value{
handle to the SimplaceWrapper object
}
\description{
Initializes the JVM and creates the SimplaceWrapper object which
is used to interact with Simplace.
}

